backend/requirements.txt
fastapi
uvicorn
motor
pydantic
bcrypt
PyJWT
python-dotenv
email-validator